topic-crawler/requirements.txt
httpx==0.26.0
bilibili-api-python==16.3
pydantic==2.8.2
uvloop==0.19.0; sys_platform != "win32"
pytest==8.3.0
